Add version selector (#5)
This adds a version selector and fixes the old code around it. The versions can be accessed by clicking on the current version which has a dropdown chevron by its name. There are also some random other fixes in this PR to some of the copy in the header/footer

Co-authored-by: driazati <driazati@users.noreply.github.com>